Add files via upload
change the original encryption method, which is based on hash encryption, to the encryption based on RSA. And, since we want to be able to encrypt an arbitrary amount of data, we use a hybrid encryption scheme. We use RSA with PKCS#1 OAEP for asymmetric encryption of an AES session key. The session key can then be used to encrypt all the actual data.